creation of custom circuits moved to the context menu
