+++ UI/src/components/pages/mainPage/createCustomBlockModal.jsx
@@ -1,0 +1,78 @@
+import React, { useState } from "react";
+import { IconCloseCross } from "../../../../assets/ui-icons.jsx";
+import { showToastError } from "../../codeComponents/logger.jsx";
+import { createCustomBlock, saveCustomBlock } from "../../utils/customBlockUtils";
+
+export default function CreateCustomBlockModal({
+                                                 isOpen,
+                                                 onClose,
+                                                 nodes,
+                                                 edges,
+                                                 onCreateFromFile,
+                                                 onCreateFromCurrent,
+                                               }) {
+  const [blockName, setBlockName] = useState("");
+  const [error, setError] = useState("");
+
+  const handleCreateFromCurrent = () => {
+    if (!blockName.trim()) {
+      showToastError("Please enter a custom block name.");
+      return;
+    }
+
+    try {
+      const customBlock = createCustomBlock(nodes, edges, blockName.trim());
+      saveCustomBlock(customBlock);
+
+      setBlockName("");
+      setError("");
+      onClose();
+
+      if (onCreateFromCurrent) onCreateFromCurrent(customBlock);
+      alert(`Block "${blockName}" created successfully!`);
+    } catch (err) {
+      console.error("Error creating block:", err);
+      setError(`Error: ${err.message}`);
+    }
+  };
+
+  const handleCreateFromFile = () => {
+    onClose();
+    if (onCreateFromFile) onCreateFromFile();
+  };
+
+  if (!isOpen) return null;
+
+  return (
+    <div className="custom-block-modal">
+      <div className="modal-content">
+        <h3>Create custom block</h3>
+        <button className="close-button-custom-circuit" onClick={onClose}>
+          <IconCloseCross SVGClassName="close-custom-circuit-cross" />
+        </button>
+
+        <div className="creation-options">
+          <button className="option-button" onClick={handleCreateFromFile}>
+            From file
+          </button>
+
+          <div className="current-circuit-option">
+            <button className="option-button" onClick={handleCreateFromCurrent}>
+              From current circuit
+            </button>
+            <div className="name-input">
+              <input
+                type="text"
+                value={blockName}
+                onChange={(e) => setBlockName(e.target.value)}
+                placeholder="New custom block name"
+                required
+              />
+              {error && <p className="error-message">{error}</p>}
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+  );
+}
